+++ src/heliax/__init__.py
@@ -148,7 +148,7 @@ from .training import Trainer
 from .workspace import Workspace
 from .workspace import workspace as workspace_context
 
-__version__ = "0.6.1"
+__version__ = "0.6.2"
 
 __all__ = [
     "GELU",